Initial Database Setup (#19)
* Established initial DB connection

Edited server.js such that the server connects to the DB when started.

* Added Recipe Schema Model

As per the title. Refer to Recipe.js to see the schema details.

* Fixed Schema Error

Fixed a big in the schema that prevented the server from starting due to an invalid schema definition.

* Added Routes to retrieve/input data

Added layouts for data retrieval and input into the database. Functionality still needs to be implemented.

* Added config file

As per title.

* Added ReadMe for Database Setup

As per title.

* Fixed fetch issue

* Updated dependencies

Updated dependencies to account for the database dependencies.

* Removed Invalid Readme

Removed the read me file as the json packages were commited therefore npm i is the only necessary preliminary function.

---------

Co-authored-by: rtal831 <rtal831@aucklanduni.ac.nz>